+++ package.json
@@ -17,7 +17,6 @@
     "@types/aws-sdk": "^2.7.0",
     "@walletconnect/sign-client": "^2.11.0",
     "@web3auth/base": "^7.0.4",
-    "@web3auth/modal": "^7.0.5",
     "apexcharts": "^3.41.1",
     "autoprefixer": "10.4.14",
     "aws-sdk": "^2.1478.0",
@@ -33,6 +32,7 @@
     "jsvectormap": "^1.5.1",
     "localforage": "^1.10.0",
     "match-sorter": "^6.3.1",
+    "mongodb": "^6.7.0",
     "next": "^14.2.3",
     "postcss": "8.4.27",
     "react": "18.2.0",
