+++ public/example/sample.json
@@ -1,0 +1,30 @@
+[
+    {
+      "id": "e9876543-abcd-1234-wxyz-abcdef123456",
+      "assetName": "Sunshine Valley Solar Farm",
+      "assetDescription": "Invest in the Sunshine Valley Solar Farm, a leading renewable energy project in the heart of California.",
+      "companyName": "Sunshine Valley Energy",
+      "assetType": "EQUITY",
+      "assetClassType": "SOLAR",
+      "companyWebsite": "http://www.sunshinevalleyenergy.com",
+      "assetOverview": {
+        "developerDetails": "Sunshine Valley Energy is a California-based solar energy company focused on developing sustainable and efficient solar projects. With a commitment to environmental stewardship and technological innovation, Sunshine Valley Energy is at the forefront of the clean energy revolution.",
+        "assetOffering": "<p style=\"margin-left:0px;\">Participate in the growth of clean energy by investing in the Sunshine Valley Solar Farm. Here are some highlights of this investment opportunity:</p><ul><li><strong>Steady Returns:</strong> Investors can expect consistent quarterly income distributions, starting from the first quarter of operation.</li><li><strong>Long-term Revenue:</strong> The Sunshine Valley Solar Farm has secured long-term contracts with reliable buyers, ensuring stable revenue streams for the foreseeable future.</li><li><strong>Robust Portfolio:</strong> This solar farm consists of multiple installations strategically located across California, maximizing energy production and revenue potential.</li><li><strong>Financial Structure:</strong> Sunshine Valley Energy plans to leverage a combination of equity and debt financing to optimize the capital structure and enhance investor returns.</li></ul><p style=\"margin-left:0px;\">Internal projections indicate an attractive annual dividend yield, reflecting the promising outlook for solar energy investments in the region.</p>",
+        "target": "<p style=\"margin-left:0px;\">The Sunshine Valley Solar Farm aims to attract environmentally conscious investors seeking both financial returns and positive impact. By investing in our project, you'll be contributing to the expansion of clean energy infrastructure and reducing carbon emissions in California and beyond.</p>",
+        "proceeds": "Proceeds from this investment opportunity will be used to fund the acquisition of the solar farm, cover transaction costs, and provide working capital for ongoing operations and maintenance."
+      },
+      "assetDetails": {
+        "issuer": {
+          "about": "Sunshine Valley Energy is dedicated to advancing solar energy adoption and sustainability initiatives in California and beyond. With a proven track record of successful project development and management, Sunshine Valley Energy continues to drive positive environmental impact and economic growth through renewable energy investments.",
+          "contractRenew": "The Sunshine Valley Solar Farm has secured revenue contracts with reputable off-takers, ensuring predictable cash flows over the project's lifecycle. These contracts offer long-term stability and the potential for contract extensions, providing additional revenue certainty for investors.",
+          "maintained": "The Sunshine Valley Solar Farm benefits from proactive asset management and ongoing maintenance efforts to optimize performance and mitigate operational risks. Comprehensive insurance coverage protects against potential damages and ensures the project's resilience against unforeseen events."
+        },
+        "financialAnalysis": "The financial analysis indicates strong potential for returns, with projected revenues exceeding initial investment within five years.",
+        "team": "Our team comprises experienced professionals with a diverse skill set, including engineers, project managers, and finance experts.",
+        "dataModel": "We utilize sophisticated data models to optimize energy production and maximize returns for investors.",
+        "secondaryMarket": "Investors have the option to trade their shares on a secondary market, providing liquidity and flexibility.",
+        "riskFactor": "While the project has mitigated many risks through careful planning, factors such as regulatory changes and natural disasters may impact performance."
+      }
+    }
+  ]
+
