+++ src/middleware.ts
@@ -5,10 +5,9 @@ export default function middleware(request: NextRequest) {
   const url = request.nextUrl.clone();
   const pathname = url.pathname;
 
-  // If the request path is not for '/assets', '/whitelist' or its child paths and is not a file or API route, redirect to home
   if (
     !pathname.startsWith('/assets') &&
-    !pathname.startsWith('/whitelist') &&
+    !pathname.startsWith('/waitlist') &&
     !pathname.startsWith('/_next') &&
     !pathname.startsWith('/api')
   ) {
