+++ tailwind.config.js
@@ -18,7 +18,7 @@ export const theme = {
     bodydark: '#AEB7C0',
     bodydark1: '#DEE4EE',
     bodydark2: '#8A99AF',
-    primary: '#4034AB',
+    primary: '#48B96C',
     secondary: '#80CAEE',
     stroke: '#E2E8F0',
     gray: '#EFF4FB',
